requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
black==26.1.0
    # via -r requirements.in
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
cfgv==3.5.0
    # via pre-commit
click==8.3.1
    # via
    #   black
    #   uvicorn
cryptography==46.0.4
    # via -r requirements.in
distlib==0.4.0
    # via virtualenv
dotenv==0.9.9
    # via -r requirements.in
fastapi==0.128.6
    # via -r requirements.in
filelock==3.20.3
    # via virtualenv
flake8==7.3.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
identify==2.6.16
    # via pre-commit
idna==3.11
    # via
    #   anyio
    #   httpx
iniconfig==2.3.0
    # via pytest
isort==7.0.0
    # via -r requirements.in
librt==0.7.8
    # via mypy
mccabe==0.7.0
    # via flake8
mypy==1.19.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nodeenv==1.10.0
    # via pre-commit
packaging==26.0
    # via
    #   black
    #   pytest
pathspec==1.0.4
    # via
    #   black
    #   mypy
platformdirs==4.5.1
    # via
    #   black
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.5.1
    # via -r requirements.in
pycodestyle==2.14.0
    # via flake8
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyjwt==2.11.0
    # via -r requirements.in
pytest==9.0.2
    # via -r requirements.in
python-dotenv==1.2.1
    # via dotenv
pytokens==0.4.1
    # via black
pyyaml==6.0.3
    # via pre-commit
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.40.0
    # via -r requirements.in
virtualenv==20.36.1
    # via pre-commit
